WIP on ClickToChatButtons page.
